remove remaining methods out of example_test into example_support
