+++ crates/nu-test-support/Cargo.toml
@@ -18,6 +18,7 @@ nu-protocol = { path="../nu-protocol", version = "0.76.1"  }
 nu-glob = { path = "../nu-glob", version = "0.76.1" }
 nu-parser = { path = "../nu-parser", version = "0.76.1" }
 nu-utils = { path="../nu-utils", version = "0.76.1"  }
+itertools = "0.10.0"
 once_cell = "1.16.0"
 num-format = "0.4.3"
 which = "4.3.0"
